+++ backend/ReservationApp/constants.py
@@ -1,0 +1,15 @@
+
+MAX_CAPACITY = 30
+
+TIME_IN_PAST_ERROR = {
+    'start_time': [
+        'The \"start_time\" field should be a future time.'
+    ]
+}
+
+OVER_CAPACITY_ERROR = {
+    'capacity': [
+        f'The specified restaurant would be over capacity during that time'
+    ]
+}
+
