+++ backend/core/constants.py
@@ -1,0 +1,10 @@
+AUTHORIZATION_ERROR_NOT_LOGGED_IN = {
+    'authorization': [
+        'Unauthorized action. Please login before trying to make a reservation'
+    ]
+}
+AUTHORIZATION_ERROR_SESSION_EXPIRED = {
+    'authorization': [
+        'Your session has expired. Please log in again.'
+    ]
+}
